pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fast-cryptography"
version = "0.1.0"
description = "Fast cryptography wrapper for lazy developers"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [{name = "Alexx-coder", email = "alexx-coder@internet.ru"}]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography"
]
keywords = ["cryptography", "fast", "wrapper", "numpy"]
dependencies = [
    "cryptography>=42.0.0,<48.0.0", 
    "base58>=2.1.0",
    "pycryptodome>=3.20.0",
    "numpy>=2.0.0",
    "base45>=0.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "mypy>=1.10",
    "ruff>=0.6.0"
]


[tool.setuptools.packages.find]
where = ["src"]  
include = ["fast_cryptography*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"fast_cryptography" = ["py.typed", "*.txt"]  